requirements.txt
langchain
langchain-community
langchain-openai
gradio
PyPDF2
faiss-cpu
sentence-transformers
openai
python-dotenv
tiktoken